@@ -2,12 +2,15 @@
 namespace App\Controller;
 
 use App\Entity\Lines;
+use App\Repository\DifficultyLevelRepository;
 use App\Repository\LinesRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Serializer\SerializerInterface;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 /**
  * @Route("/api", name="api")
@@ -15,14 +18,19 @@ use Symfony\Component\Serializer\SerializerInterface;
 class LinesApiController extends AbstractController
 {
 
-    private $linesRepository;
+    private LinesRepository $linesRepository;
 
-    private $entityManager;
+    private EntityManagerInterface $entityManager;
 
-    public function __construct(LinesRepository $linesRepository, EntityManagerInterface $entityManager)
+    private DifficultyLevelRepository $difficultyLevelRepository;
+
+    private ValidatorInterface $validator;
+
+    public function __construct(LinesRepository $linesRepository, EntityManagerInterface $entityManager, DifficultyLevelRepository $difficultyLevelRepository, ValidatorInterface $validator)
     {
         $this->linesRepository = $linesRepository;
         $this->entityManager = $entityManager;
+        $this->difficultyLevelRepository = $difficultyLevelRepository;
     }
 
     /**
@@ -41,6 +49,40 @@ class LinesApiController extends AbstractController
     public function show(Lines $line): Response
     {
         return $this->json($line, 200, [], ['groups' => 'show_line']);
+    }
+
+    /**
+     * @Route("/lines", name="api_create_lines", methods={"POST"})
+     */
+    public function create(Request $request, SerializerInterface $serializer): Response
+    {
+
+        $data = $serializer->deserialize($request->getContent(), Lines::class, 'json');
+
+        try {
+            $line = new Lines();
+            $difficulties = $data->getDifficulties();
+
+            foreach ($difficulties as $difficulty)
+            {
+                $line->addDifficulty($this->difficultyLevelRepository->findOneBy(['name' => $difficulty->getName()]));
+            }
+
+            $this->entityManager->persist($line);
+            $this->entityManager->flush();
+            return $this->json('OK', 200, []);
+
+
+        } catch(\Exception $e){
+            $errorMessage = $e->getMessage();
+            return $this->json(['message' => $errorMessage], 400, []);
+        }
+
+
+
+
+
+
     }
 
 }
